rename list ask api
