--- venus-shared/types/market/retrieval_ask.go
@@ -1,15 +1,0 @@
-package market
-
-import (
-	"github.com/filecoin-project/go-address"
-	"github.com/filecoin-project/go-state-types/abi"
-)
-
-type RetrievalAsk struct {
-	Miner                   address.Address
-	PricePerByte            abi.TokenAmount
-	UnsealPrice             abi.TokenAmount
-	PaymentInterval         uint64
-	PaymentIntervalIncrease uint64
-	TimeStamp
-}
