+++ venus-shared/types/market/storage.go
@@ -3,17 +3,77 @@ package market
 import (
 	"time"
 
+	datatransfer "github.com/filecoin-project/go-data-transfer"
+	"github.com/filecoin-project/go-fil-markets/filestore"
 	"github.com/filecoin-project/go-fil-markets/retrievalmarket"
-	"github.com/filecoin-project/venus/venus-shared/types"
-
 	"github.com/filecoin-project/go-fil-markets/storagemarket"
-
 	"github.com/filecoin-project/go-state-types/abi"
-	"github.com/filecoin-project/go-state-types/crypto"
+	crypto "github.com/filecoin-project/go-state-types/crypto"
+	"github.com/filecoin-project/venus/venus-shared/types"
 	"github.com/ipfs/go-cid"
+	"github.com/libp2p/go-libp2p/core/peer"
+	cbg "github.com/whyrusleeping/cbor-gen"
 )
 
-// todo  move to sealer
+type MinerDeal struct {
+	types.ClientDealProposal
+	ProposalCid           cid.Cid
+	AddFundsCid           *cid.Cid
+	PublishCid            *cid.Cid
+	Miner                 peer.ID
+	Client                peer.ID
+	State                 storagemarket.StorageDealStatus
+	PiecePath             filestore.Path
+	PayloadSize           uint64
+	MetadataPath          filestore.Path
+	SlashEpoch            abi.ChainEpoch
+	FastRetrieval         bool
+	Message               string
+	FundsReserved         abi.TokenAmount
+	Ref                   *storagemarket.DataRef
+	AvailableForRetrieval bool
+
+	DealID       abi.DealID
+	CreationTime cbg.CborTime
+
+	TransferChannelID *datatransfer.ChannelID `json:"TransferChannelId"`
+	SectorNumber      abi.SectorNumber
+
+	Offset      abi.PaddedPieceSize
+	PieceStatus PieceStatus
+
+	InboundCAR string
+
+	TimeStamp
+}
+
+func (deal *MinerDeal) FilMarketMinerDeal() *storagemarket.MinerDeal {
+	return &storagemarket.MinerDeal{
+		ClientDealProposal:    deal.ClientDealProposal,
+		ProposalCid:           deal.ProposalCid,
+		AddFundsCid:           deal.AddFundsCid,
+		PublishCid:            deal.PublishCid,
+		Miner:                 deal.Miner,
+		Client:                deal.Client,
+		State:                 deal.State,
+		PiecePath:             deal.PiecePath,
+		MetadataPath:          deal.MetadataPath,
+		SlashEpoch:            deal.SlashEpoch,
+		FastRetrieval:         deal.FastRetrieval,
+		Message:               deal.Message,
+		FundsReserved:         deal.FundsReserved,
+		Ref:                   deal.Ref,
+		AvailableForRetrieval: deal.AvailableForRetrieval,
+
+		DealID:       deal.DealID,
+		CreationTime: deal.CreationTime,
+
+		TransferChannelId: deal.TransferChannelID,
+		SectorNumber:      deal.SectorNumber,
+
+		InboundCAR: deal.InboundCAR,
+	}
+}
 
 // PendingDealInfo has info about pending deals and when they are due to be
 // published
@@ -22,102 +82,6 @@ type PendingDealInfo struct {
 	PublishPeriodStart time.Time
 	PublishPeriod      time.Duration
 }
-
-type SectorOffset struct {
-	Sector abi.SectorNumber
-	Offset abi.PaddedPieceSize
-}
-
-// DealInfo is a tuple of deal identity and its schedule
-type PieceDealInfo struct {
-	PublishCid   *cid.Cid
-	DealID       abi.DealID
-	DealProposal *types.DealProposal
-	DealSchedule DealSchedule
-	KeepUnsealed bool
-}
-
-// DealSchedule communicates the time interval of a piecestorage deal. The deal must
-// appear in a sealed (proven) sector no later than StartEpoch, otherwise it
-// is invalid.
-type DealSchedule struct {
-	StartEpoch abi.ChainEpoch
-	EndEpoch   abi.ChainEpoch
-}
-
-type SectorState string
-
-type SealTicket struct {
-	Value abi.SealRandomness
-	Epoch abi.ChainEpoch
-}
-
-type SealSeed struct {
-	Value abi.InteractiveSealRandomness
-	Epoch abi.ChainEpoch
-}
-
-type SectorLog struct {
-	Kind      string
-	Timestamp uint64
-
-	Trace string
-
-	Message string
-}
-
-type SectorInfo struct {
-	SectorID     abi.SectorNumber
-	State        SectorState
-	CommD        *cid.Cid
-	CommR        *cid.Cid
-	Proof        []byte
-	Deals        []abi.DealID
-	Ticket       SealTicket
-	Seed         SealSeed
-	PreCommitMsg *cid.Cid
-	CommitMsg    *cid.Cid
-	Retries      uint64
-	ToUpgrade    bool
-
-	LastErr string
-
-	Log []SectorLog
-
-	// On Chain Info
-	SealProof          abi.RegisteredSealProof // The seal proof type implies the PoSt proof/s
-	Activation         abi.ChainEpoch          // Epoch during which the sector proof was accepted
-	Expiration         abi.ChainEpoch          // Epoch during which the sector expires
-	DealWeight         abi.DealWeight          // Integral of active deals over sector lifetime
-	VerifiedDealWeight abi.DealWeight          // Integral of active verified deals over sector lifetime
-	InitialPledge      abi.TokenAmount         // Pledge collected to commit this sector
-	// Expiration Info
-	OnTime abi.ChainEpoch
-	// non-zero if sector is faulty, epoch at which it will be permanently
-	// removed if it doesn't recover
-	Early abi.ChainEpoch
-}
-
-type SealedRef struct {
-	SectorID abi.SectorNumber
-	Offset   abi.PaddedPieceSize
-	Size     abi.UnpaddedPieceSize
-}
-
-type SealedRefs struct {
-	Refs []SealedRef
-}
-
-type AddrUse int
-
-const (
-	PreCommitAddr AddrUse = iota
-	CommitAddr
-	DealPublishAddr
-	PoStAddr
-
-	TerminateSectorsAddr
-)
 
 // StorageDealStatistic storage  statistical information
 // The struct is used here for statistical information that may need to be added in the future
@@ -131,12 +95,7 @@ type RetrievalDealStatistic struct {
 	DealsStatus map[retrievalmarket.DealStatus]int64
 }
 
-type StorageStatus struct {
-	Capacity  int64
-	Available int64
-	Reserved  int64
-}
-
+// SignedStorageAsk use to record provider's requirement in database
 type SignedStorageAsk struct {
 	Ask       *storagemarket.StorageAsk
 	Signature *crypto.Signature
@@ -148,8 +107,4 @@ func (sa *SignedStorageAsk) ToChainAsk() *storagemarket.SignedStorageAsk {
 		Ask:       sa.Ask,
 		Signature: sa.Signature,
 	}
-}
-
-func FromChainAsk(s *storagemarket.SignedStorageAsk) *SignedStorageAsk {
-	return &SignedStorageAsk{Ask: s.Ask, Signature: s.Signature}
 }
